Moved GraqlDelete, GraqlInsert and Pattern tests to their respective package in Graql
