+++ graql/java/query/test/GraqlInsertTest.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package grakn.core.graql.query;
+package graql.lang.query.test;
 
 import graql.lang.Graql;
 import graql.lang.query.GraqlInsert;
